Move DonationFormContactInfoPage to pages folder
