+++ client/src/components/organisms/style/index.scss
@@ -14,4 +14,3 @@
 @import "FormModal";
 @import "LogoModal";
 @import "DeleteRequestTypeDialog";
-@import "DonationFormContactInfoPage";
