+++ client/src/pages/style/index.scss
@@ -1,4 +1,5 @@
 @import "AdminHomepage";
+@import "DonationFormContactInfoPage";
 @import "DonorHomepage";
 @import "ResetPasswordModal";
 @import "SendResetPasswordEmailModal";
